Split code to more files. Mainly move processor code to its own file.
